Move each stories to stories folder and bundle button, modal to components
