+++ client/app/stories/category/CategoryItem.stories.tsx
@@ -2,7 +2,7 @@ import { fn } from '@storybook/test'
 import { CategorySection } from '@/app/ui'
 import { colors } from '@/app/styles'
 import type { Meta, StoryObj } from '@storybook/react'
-import { CategoryItem } from '../CategoryItem'
+import { CategoryItem } from '../../ui/category/CategoryItem'
 
 const categoryItem = {
   title: 'Example/category',
