+++ client/app/stories/category/CategorySection.stories.tsx
