+++ client/app/stories/components/Button.stories.ts
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { fn } from '@storybook/test'
 import { ButtonSize, ButtonsTheme } from '@/app/types'
-import { Button } from '..'
+import { Button } from '../../ui/common'
 
 const meta = {
-  title: 'Example/Button',
+  title: 'Example/Components/Button',
   component: Button,
   parameters: {
     layout: 'centered',
